create tag endpoint_file
