+++ app/drivers/api/endpoints/event.py
@@ -3,7 +3,6 @@ from typing import Optional
 import app.domains.entities as entities
 import app.usecases as usecases
 from app.drivers.api.deps import get_event_usecase
-import app.domains.entities.tag as tag_entities
 
 
 router = APIRouter()
@@ -56,27 +55,6 @@ async def get_by_id(
     return get_event
 
 
-@router.post("/{event_id}/tag", response_model=tag_entities.Tag)
-async def create_event_tag(
-    *,
-    tag_in: tag_entities.TagCreate,
-    eu: usecases.EventUsecase = Depends(get_event_usecase)
-) -> tag_entities.Tag:
-    return eu.create_tag(obj_in=tag_in)
-
-
-@router.get("/{event_id}/tag/{tag_id}", response_model=tag_entities.Tag)
-async def get_tag_by_id(
-    event_id: int, tag_id: int, eu: usecases.EventUsecase = Depends(get_event_usecase)
-) -> Optional[tag_entities.Tag]:
-    get_event: Optional[tag_entities.Tag] = eu.get_tag_by_id(
-        event_id=event_id, tag_id=tag_id
-    )
-    if get_event is None:
-        raise HTTPException(status_code=404)
-    return get_event
-
-
 # todo: 不要
 @router.get("/", response_model=entities.ListEventsResponse)
 async def get_events(
@@ -98,18 +76,3 @@ async def delete_event(
     if deleted_event is None:
         raise HTTPException(status_code=404, detail="イベント削除エラー")
     return deleted_event
-
-
-@router.delete("/{event_id}/tag/{tag_id}", response_model=tag_entities.Tag)
-async def delete_tag(
-    *,
-    event_id: int,
-    tag_id: int,
-    eu: usecases.EventUsecase = Depends(get_event_usecase)
-) -> Optional[tag_entities.Tag]:
-    deleted_tag: Optional[tag_entities.Tag] = eu.delete_tag(
-        event_id=event_id, tag_id=tag_id
-    )
-    if deleted_tag is None:
-        raise HTTPException(status_code=404, detail="イベント削除エラー")
-    return deleted_tag
